moved SetResult API to API package
