+++ daas/daas_app/api/internals/__init__.py
@@ -1,0 +1,1 @@
+from .set_result import SetResultApiView
