+++ daas/daas_app/urls.py
@@ -56,7 +56,7 @@ urlpatterns = [
     re_path(r'^api/get_token/?$', api.get_token_view, name='api_get_token'),
 
     # Private API (only reachable within the docker network)
-    re_path(r'^internal/api/set_result/?$', views.SetResult.as_view(), name='set_result'),
+    re_path(r'^internal/api/set_result/?$', api.internals.SetResultApiView.as_view(), name='set_result'),
 ]
 
 urlpatterns = format_suffix_patterns(urlpatterns)
